requirements.txt
cffi==1.14.5
GitPython==3.1.14
numpy==1.20.1
scipy==1.6.1
simplejson==3.16.0
